+++ example/lib/pages/web/web_home_page.dart
@@ -5,10 +5,9 @@ import '../../widgets/calendar_configs.dart';
 import '../../widgets/calendar_views.dart';
 
 class WebHomePage extends StatefulWidget {
-  WebHomePage({this.selectedView = CalendarView.month, this.onThemeChange});
+  WebHomePage({this.selectedView = CalendarView.month});
 
   final CalendarView selectedView;
-  final void Function(bool)? onThemeChange;
 
   @override
   _WebHomePageState createState() => _WebHomePageState();
@@ -34,7 +33,6 @@ class _WebHomePageState extends State<WebHomePage> {
             child: CalendarConfig(
               onViewChange: _setView,
               currentView: _selectedView,
-              onThemeChange: widget.onThemeChange,
             ),
           ),
           Expanded(