+++ lib/src/theme/day_view_theme_data.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'dark_app_colors.dart';
-import 'light_app_colors.dart';
+import 'calendar_view_colors.dart';
 
 class DayViewThemeData extends ThemeExtension<DayViewThemeData> {
   /// Define custom colors
@@ -36,27 +35,27 @@ class DayViewThemeData extends ThemeExtension<DayViewThemeData> {
 
   /// Get pre-defined colors for light theme
   DayViewThemeData.light()
-      : hourLineColor = LightAppColors.surfaceContainerHighest,
-        halfHourLineColor = LightAppColors.surfaceContainerHighest,
-        quarterHourLineColor = LightAppColors.surfaceContainerHighest,
-        pageBackgroundColor = LightAppColors.surfaceContainerLowest,
-        liveIndicatorColor = LightAppColors.primary,
-        headerIconColor = LightAppColors.onPrimary,
-        headerTextColor = LightAppColors.onPrimary,
-        headerBackgroundColor = LightAppColors.primary,
-        timelineTextColor = LightAppColors.onSurface;
+      : hourLineColor = CalendarViewColors.light.surfaceContainerHighest,
+        halfHourLineColor = CalendarViewColors.light.surfaceContainerHighest,
+        quarterHourLineColor = CalendarViewColors.light.surfaceContainerHighest,
+        pageBackgroundColor = CalendarViewColors.light.surfaceContainerLowest,
+        liveIndicatorColor = CalendarViewColors.light.primary,
+        headerIconColor = CalendarViewColors.light.onPrimary,
+        headerTextColor = CalendarViewColors.light.onPrimary,
+        headerBackgroundColor = CalendarViewColors.light.primary,
+        timelineTextColor = CalendarViewColors.light.onSurface;
 
   /// Get pre-defined colors for dark theme
   DayViewThemeData.dark()
-      : hourLineColor = DarkAppColors.surfaceContainerHighest,
-        halfHourLineColor = DarkAppColors.surfaceContainerHighest,
-        quarterHourLineColor = DarkAppColors.surfaceContainerHighest,
-        pageBackgroundColor = DarkAppColors.surfaceContainerLowest,
-        liveIndicatorColor = DarkAppColors.primary,
-        headerIconColor = DarkAppColors.onPrimary,
-        headerTextColor = DarkAppColors.onPrimary,
-        headerBackgroundColor = DarkAppColors.primary,
-        timelineTextColor = DarkAppColors.onSurface;
+      : hourLineColor = CalendarViewColors.dark.surfaceContainerHighest,
+        halfHourLineColor = CalendarViewColors.dark.surfaceContainerHighest,
+        quarterHourLineColor = CalendarViewColors.dark.surfaceContainerHighest,
+        pageBackgroundColor = CalendarViewColors.dark.surfaceContainerLowest,
+        liveIndicatorColor = CalendarViewColors.dark.primary,
+        headerIconColor = CalendarViewColors.dark.onPrimary,
+        headerTextColor = CalendarViewColors.dark.onPrimary,
+        headerBackgroundColor = CalendarViewColors.dark.primary,
+        timelineTextColor = CalendarViewColors.dark.onSurface;
 
   @override
   ThemeExtension<DayViewThemeData> copyWith({